Moved clases to approriate structure
