+++ src/test/java/uk/ac/ebi/ampt2d/test/persistence/TestMonotonicEntity.java
@@ -17,7 +17,7 @@
  */
 package uk.ac.ebi.ampt2d.test.persistence;
 
-import uk.ac.ebi.ampt2d.commons.generators.ModelHashAccession;
+import uk.ac.ebi.ampt2d.commons.accession.generators.ModelHashAccession;
 import uk.ac.ebi.ampt2d.test.TestModel;
 
 import javax.persistence.Column;
